Fixed middleware to work
