+++ .gitignore
@@ -1,5 +1,5 @@
 .idea
-handlers/
+dev/
 main.go
 GoWeb
 Makefile
